+++ Changelog.md
@@ -33,6 +33,7 @@
 - More robust slicing!
 - Fix crash when slicing to different resolutions in the same session
 - Lag free slice preview scrubbing (layer data is decompressed on the gpu)
+- Add buttons to reload or replace a model
 
 ## v0.8.0 &mdash; June 30th, 2026